This adds support for running dnscontrol-action in Forgejo Actions.
Tested on Codeberg but any Forgejo instance should be compatible. No
changes are needed on the user side apart from the `uses` argument in
the step definition needs to reference the full URL to the project (this
applies only to Forgejo Actions users).

Usage in Forgejo Actions:

```yaml
- name: DNSControl preview
  uses: https://github.com/DNSControl/dnscontrol-action
  with:
    cmdargs: preview
```

Changes:

* Forgejo Actions support
* Downloading releases is more reliable in GHA and Forgejo Actions
* The action is now self contained, no external actions called
* Run shellcheck on files matching `bin/*sh`
* Better debugging capability to assist with testing. To get debug
output, run with `DEBUG_ACTION: true` set in an `env` block.

+++ bin/shellcheck.sh
@@ -2,9 +2,11 @@
 
 set -eo pipefail
 
-# shellcheck wrapper - to be run locally and in PR workflow:
-# extract each bash run step from action.yml and all workflows and run them
-# through shellcheck for safety. requires yq and shellcheck in $PATH
+# This script checks all project shell scripts for common errors and style
+# issues. It extracts each bash run step from action.yml and all workflows and
+# runs them through shellcheck for safety. requires yq and shellcheck in $PATH
+
+# Run locally with: [bash|zsh] bin/shellcheck.sh
 
 FAILS=0
 IFS="
@@ -12,50 +14,61 @@ IFS="
 
 # Create an array of YAML files (recent bash and zsh supported here)
 if [[ -n "$BASH_VERSION" ]]; then
-  YAMLFILES=()
-  while read -r YAMLFILE; do
-      YAMLFILES+=("$YAMLFILE")
-  done < <(find action.yml .github/workflows -type f -name \*yml -print -o -name \*yaml -print)
+  scripts=()
+  while read -r script; do
+      scripts+=("$script")
+  done < <(find action.yml .github/workflows bin -type f -name \*sh -print -o -name \*yml -print -o -name \*yaml -print)
 else
   if [[ -n "$ZSH_VERSION" ]]; then
-    YAMLFILES=("${(@f)$(find action.yml .github/workflows -type f -name \*yml -print -o -name \*yaml -print)}")
+    # shellcheck disable=SC2296
+    scripts=("${(@f)$(find action.yml .github/workflows bin -type f -name \*sh -print -o -name \*yml -print -o -name \*yaml -print)}")
   else
     echo "This script requires a recent bash or zsh shell"
     exit 1
   fi
 fi
 
+errors=""
 
-for YAMLFILE in "${YAMLFILES[@]}"; do
+for script in "${scripts[@]}"; do
 
-  echo "Checking $YAMLFILE"
+  echo "Checking $script"
 
-  # find bash run steps in workflow jobs.*.steps[] and action runs.steps[] 
-  for n in $(yq '.. | select(has("steps")) | .steps[] | select(.shell == "bash") | .name' < "$YAMLFILE")
-  do
-    # do not error out while in this inner loop: we want to find all problems
+  if [[ "${script##*.}" == "sh" ]]; then
     set +e
-
-    # prepare a safe location to create a temporary script for evaluation
-    tmpscript=$(mktemp)
-
-    # extract the *.steps[].run elements into the temp file 
-    yq ".. | select(has(\"steps\")) | .steps[] | select(.name == \"$n\") | .run" < "$YAMLFILE" > "$tmpscript"
-
-    # run the checks
-    shellcheck --shell bash -S warning "$tmpscript"
-    if [ $? -ne 0 ]; then
-      echo ""
-      echo "The script in $YAMLFILE step \"$n\" did not pass shellcheck"
-      echo ""
+    if ! shellcheck --shell bash -S warning "$script"; then
+      errors="$errors\n$script did not pass shellcheck"
       ((FAILS++))
     fi
-    rm "$tmpscript"
     set -e
-  done
+  fi
+
+  if [[ "${script##*.}" == "yml" ]] || [[ "${script#*.}" == "yaml" ]]; then
+    # find bash run steps in workflow jobs.*.steps[] and action runs.steps[]
+    for n in $(yq '.. | select(has("steps")) | .steps[] | select(.shell == "bash") | .name' < "$script")
+    do
+      # do not error out while in this inner loop: we want to find all problems
+      set +e
+
+      # prepare a safe location to create a temporary script for evaluation
+      tmpscript=$(mktemp)
+
+      # extract the *.steps[].run elements into the temp file
+      yq ".. | select(has(\"steps\")) | .steps[] | select(.name == \"$n\") | .run" < "$script" > "$tmpscript"
+
+      # run the checks
+      if ! shellcheck --shell bash -S warning "$tmpscript"; then
+        errors="$errors\n$script step \"$n\" did not pass shellcheck"
+        ((FAILS++))
+      fi
+      rm "$tmpscript"
+      set -e
+    done
+  fi
 done
 
 if [[ $FAILS -ne 0 ]]; then
   echo "$FAILS error(s) or warning(s) detected, please fix and run bin/shellcheck.sh again"
+  echo "issues:$errors"
   exit 1
 fi
